잠시

+++ dacoach/src/main/java/com/dacoach/mapper/admin/company/AdminCompanyMapper.java
@@ -13,19 +13,19 @@ public interface AdminCompanyMapper {
 
     // 기업 회원 관리
     List<Map<String, Object>> companyList();
-    Map<String, Object> companyDetail(@Param("usersIdx") long usersIdx);
-    String getCertFilePath(@Param("userIdx") long userIdx);
+    Map<String, Object> companyDetail(@Param("usersIdx") int usersIdx);
+    String getCertFilePath(@Param("userIdx") int userIdx);
     int insertCert(CertDTO dto);
     int hasCert(@Param("userIdx") int userIdx);
     
     int countCompanyTotal();    
-    int updateUserStatus(@Param("userIdx") long userIdx, @Param("status") String status);
+    int updateUserStatus(@Param("userIdx") int userIdx, @Param("status") String status);
     int insertEmbeddedHistory(Map<String, Object> param);
-    int closeLatestEmbeddedHistory(@Param("userIdx") long userIdx);
+    int closeLatestEmbeddedHistory(@Param("userIdx") int userIdx);
     
-    int countCertByUserAndType(@Param("userIdx") long userIdx, @Param("certType") String certType);
-    int updateCertStatus(@Param("user_idx") long userIdx, @Param("cert_status") String certStatus);
-    int insertCertStatus(@Param("user_idx") long userIdx, @Param("cert_status") String certStatus);
+    int countCertByUserAndType(@Param("userIdx") int userIdx, @Param("certType") String certType);
+    int updateCertStatus(@Param("user_idx") int userIdx, @Param("cert_status") String certStatus);
+    int insertCertStatus(@Param("user_idx") int userIdx, @Param("cert_status") String certStatus);
     
     // 클래스 목록
     List<Map<String, Object>> selectClassPage(Map<String, Object> param);
